cbus: support libev-less cpipe
cpipe uses libev only for auto-flush. That is, when a first
message is pushed, the pipe would schedule its own flush to the
end of the current iteration of the event loop. This allows
automatic batching, when multiple fibers during one iteration of
the loop would push stuff into the pipe.

The feature is quite insignificant though. It can easily be turned
off and cpipe becomes just a unidirectional mutex-locked queue. It
can still be used like before then, but won't have an auto-flush.

The commit enables exactly that. It is a preparation for the new
public API which would allow users to push stuff into TX thread
from other threads using per-thread cpipes without libev.

NO_DOC=internal
NO_CHANGELOG=internal