fuzz: populate engine fuzzer with fixed-size types
The commit populates engine fuzzer with all Tarantool fixed-size types.
These types are often used in `memcs` so we should support them since
we are going to test it as well. Also, it might be useful for other
engines.

Along the way, add operation '+' to type `double` - it seems to be
forgotten.

Part of tarantool/tarantool-ee#834

NO_CHANGELOG=fuzzing
NO_DOC=fuzzing