fuzz: support memcs engine in engine fuzzing
The commit adds `memcs` support to our engine fuzzing - this engine has
some severe limitations, so the commit makes the test respect them.
Also, it populates the test with new memcs-specific ops:

1. `COLUMN_SCAN_OP` - iterates over values in specified columns. Size of
   each batch is chosen randomly, also a fiber yield is randomly
   injected after every read.
2. `ARROW_STREAM_OP` - fetches data from specified columns in Apache Arrow
   format. All batches have the same size, it is randomly chosen at the
   start.
3. `INSERT_ARROW_OP` - randomly chooses columns and generates equally
   sized batches for each column, then inserts them as a batch in Apache
   Arrow format.

Also, `memcs` does not support some operations like upsert - the commit
marks them as vinyl-and-memtx-specific. The same goes for type
`integer`, so the commit adds a new attribute `engines` to the list of
Tarantool types and marks `integer` as vinyl-and-memtx-specific.

We have `space:insert_arrow` function that accepts a string with Apache
Arrow data. The commit doesn't cover it since it's actually the same as
a Lua wrapper that we use except for an additional decoding step.
Apache Arrow decoder should be fuzzed in a separate test and have
nothing in common with Tarantool engines, so we don't need to cover it
here.

Part of tarantool/tarantool-ee#834

NO_CHANGELOG=fuzzing
NO_DOC=fuzzing