fuzz: support memcs engine in engine fuzzing
The commit adds `memcs` support to our engine fuzzing - this engine has
some severe limitations, so the commit makes the test respect them.
Also, it populates the test with new memcs-specific ops:

1. `COLUMN_SCAN_OP` - iterates over values in specified columns. Size of
   each batch is chosen randomly, also a fiber yield is randomly
   injected after every read.
2. `ARROW_STREAM_OP` - fetches data from specified columns in Apache Arrow
   format. All batches have the same size, it is randomly chosen at the
   start.
3. `INSERT_ARROW_OP` - randomly chooses columns and generates equally
   sized batches for each column, then inserts them as a batch in Apache
   Arrow format.

Since there is no machinery to dynamically update the list of operations
`ops`, let's simply call abovementioned ops for all engines, but return
early if the engine is not `memcs`.

We have `space:insert_arrow` function that accepts a string with Apache
Arrow data. The commit doesn't cover it since it's actually the same as
a Lua wrapper that we use except for an additional decoding step.
Apache Arrow decoder should be fuzzed in a separate test and have
nothing in common with Tarantool engines, so we don't need to cover it
here.

Part of tarantool/tarantool-ee#834

NO_CHANGELOG=fuzzing
NO_DOC=fuzzing