box: make bootstrap_leader cfg option dynamic
bootstrap_leader configuration option only takes effect during instance
bootstrap. For this reason, the option was made static. One can't change
the bootstrap_leader after initial box.cfg() call.

So when a user has bootstrap_strategy = 'config' and joins new nodes
during replicaset lifespan, he ends up with instances having different
`bootstrap_leader` configuration, depending on which node was writable
at the moment of each instance bootstrap.

It would be better to have matching configuration on each replica
set member with no exceptions and to be able to set bootstrap_leader to
the same value on all the instances of the replica set, even on the ones
which are already bootstrapped.

Let's allow changing `bootstrap_leader` option even though it affects
nothing.

While we're at it, fix a memory leak of bootstrap leader uri on
tarantool shutdown.

Closes #10604

@TarantoolBot document
Title: clarify `bootstrap_leader` configuration option meaning

The option `bootstrap_leader` is dynamic, but it's important to note
that changing it after instance bootstrap affects nothing. It's left
dynamic simply for the sake of centralized configuration, when it's
preferable to have matching configurations on all instances of a replica
set.