ci: fixate the runner for perf_micro workflow
This patch fixates the GitHub CI runner with the label 'performance' for
the aforementioned workflow, if it is run from the master, release
branch, or triggered by PR from the main repository. It helps to avoid
disturbance in the results of benchmarks due to different machine
parameters.

NO_TEST=ci
NO_CHANGELOG=ci
NO_DOC=ci