config: add replication.disabled option
The commit adds the main logic behind the `replication.disabled` option.

One bullet from #10776 is not solved here: refuse to bootstrap if the
option is set. This functionality is to be implemented in a separate
patch.

Part of #10776

@TarantoolBot document
Title: config: replication.disabled option

Tarantool 3.3 adds the new YAML configuration option:
`replication.disabled`. It is useful to temporary exclude an instance
from its replica set.

The instance with `replication.disabled = true` disconnects from all the
upstreams.

All the other instances in the replica set refuse to fetch data from the
disabled instance.

If the instance is non-anonymous it is kept in the `_cluster` system
space, so the replication from/to it can be enabled back.

See details in https://github.com/tarantool/tarantool/issues/10776.