replication: drop dead code in replica_set_id
Function `replica_set_id` registers a WAL GC consumer for anonymous
replicas on their transition to regular ones. The problem is that
anonymous replicas never appear there since Tarantool doesn't store them
when they are disconnected, and in order to register, they mustn't be
connected (except for registration connection itself, but it doesn't
creates a replica object), so Tarantool creates a new replica on
registration that is not anonymous by default.

Moreover, we already create a consumer in `box_process_register` and
check if replica had no other with an assertion. The dead snippet does
the same thing, so one of assertions would fail if the code was actually
executed. Since we are going to keep disconnected anonymous replicas,
this problem will get in the way, so let's simply drop the snippet.

Part of #10755

NO_TEST=drop dead code
NO_CHANGELOG=drop dead code
NO_DOC=drop dead code