test: add cluster config consistency test cases
If an option is added to or removed from the cluster config schema, or a
default value is changed, these test cases should catch the change. It
gives some protection from an occasional modification.

We have such a check for the instance config schema. This patch
implements something similar for the cluster configuration schema.

The test cases verify a list of fields on each of the levels: global,
group, replicaset, instance.

Made in the scope of #10796.

NO_DOC=no functional changes
NO_CHANGELOG=see NO_DOC