replication: provide anonymous replicas with WAL GC consumers
WAL GC consumers are created for non-anonymous replicas to prevent
required xlogs from being deleted so that replicas won't fall behind
because of replication lag. The commit makes Tarantool register WAL GC
consumers for anonymous replica as well.

In order to recognize replica on IPROTO_FETCH_SNAPSHOT, it was populated
with optional IPROTO_INSTANCE_UUID key - if it passed, replica object is
created if it's needed and populated with WAL GC consumer. Regarding
IPROTO_SUBSCRIBE, we always create WAL GC consumers for anonymous replicas.
What's about Tarantool implementation of anonymous replica, the commit
makes it send IPROTO_INSTANCE_UUID in IPROTO_FETCH_SNAPSHOT so now it
retains xlogs right from the beginning.

Since disconnected anonymous replica is actually orphan, it will be
deleted at the first chance along with its WAL GC consumer. To prevent
that from happening, any anonymous replica is considered as not orphan
since it has lifetime different from regular replicas.

Important to note that the commit changes the structure of REGISTER and
JOIN endpoints. Before the commit they used the following steps:
1. Retrieve the replica object and clear its WAL GC state, if any.
2. Register a WAL GC consumer but not put it to the replica object.
3. After successful registration, put the created consumer to the replica.
The problem here is if the registration fails, replica loses its
in-memory consumer, so we will never reach if from anonymous replica
garbage collector and its persistent WAL GC consumer will linger in the
system space. To solve this problem, let's register the new WAL GC consumer
and put it to the replica right after deleting the old one. If there is no
corresponding replica object in memory, let's create an anonymous one.
So actual behavior change is that any replica is actually registered as
an anonymous one before the registration. The good thing is we are going
to implement a garbage collector for disconnected anonymous replicas, so
failed to register replicas will be deleted eventually.

Since the commit introduces a helper for connecting replicas, let's
enhance protection against duplicatate connection with the same UUID.
Currently when a replica connects we check if relay is not connected,
that's our criteria of a duplicate connection. However, sometimes
replica can be connected and not using relay (for example, it is
persisting WAL GC consumer or is in initial join stage that doesn't use
replica's relay) - in this case a duplicate connection can actually
happen and break Tarantool. The commit introduces a new guard to handle
disconnection automatically. By the way, we somewhy didn't check for a
duplicate connection on JOIN at all (seems we just forgot to do it).

Since connected replica is not orphan actually and WAL GC consumer can
be used not only in relay but also in replication endpoints, the commit
introduces new callblack `replica_on_disconnect` - it removes WAL GC
state of replicas evicted from the cluster and deletes orphan ones,
previously we did it in `replica_on_relay_stop`.

Part of #10755

NO_CHANGELOG=later
NO_DOC=later