config: add failover learners section
This patch adds the support for
`failover.replicasets.<replicaset-name>.learners` configuration section
and corresponding checks for it.

This option is used to specify the list of instances from the specific
replicaset that can't be selected as a master by the supervised failover
coordinator.

The supervised coordinator is an enterprise-edition feature.

EE issue: tarantool/tarantool-ee#991
EE PR: tarantool/tarantool-ee#992

NO_DOC=documented in EE PR
NO_CHANGELOG=changelog in EE PR