replication: anon replicas don't participate in elections
Fixed a bug where anonymous replicas could participate in elections or even
be chosen as a leader. It is now forbidden to configure a replica so
that `replication_anon` is `true` and `election_mode` is not `off`.
It is also now prohibited to issue PROMOTE from an anonymous replica.

Closes #10561

@TarantoolBot document
Title: changes in `replication_anon` and `election_mode` configuration
Product: Tarantool
Since: 3.3

[replication-anon] https://www.tarantool.io/ru/doc/latest/reference/configuration/#cfg-replication-replication-anon

```diff
- In order to make a replica anonymous, pass the option `replication_anon=true` to `box.cfg` and set `read_only` to `true`.
+ In order to make a replica anonymous, pass the option `replication_anon=true` to `box.cfg`, set `read_only` to `true` and `election_mode` to `off`.
```

[election_mode] https://www.tarantool.io/ru/doc/latest/reference/configuration/#cfg-replication-election-mode

For an anonymous replica, `election_mode` can only be set to `off`.