memtx: reserve extents in the RTREE index itself
We need to reserve extents before a writing operation on the RTREE
because the data structure does not handle memory allocation errors.
Let's move the reservation from the memtx space into the RTREE index
itself because it is not required for other indexes anymore.

NO_DOC=internal
NO_CHANGELOG=internal