config: don't replicate from isolated instance
The isolated instance goes to RO and drops iproto connections; thus, the
replication from it is effectively stopped. There is no sense to retry
an attempt to connect to it and flood logs with errors. Let's just
remove the isolated instance from the upstreams list.

Part of #10796

NO_DOC=tarantool/doc#4632
NO_CHANGELOG=added together with the configuration option