memtx: do not abort unrelated transactions on DDL
Currently we simply abort all concurrent transactions on any DDL. We've
already done a few workarounds for these problems - filter that prevents
DDL on temporary spaces from aborting fully remote transactions and vice
versa, and an engine flag `ENGINE_TXM_HANDLES_DDL` to prevent memtx DDL
from aborting vinyl transactions. The commit removes above-mentioned
temporary solutions and makes memtx abort only related transactions on DDL.

The new approach is to abort all writers and readers when a space is
invalidated (it happens on replace in space cache). To abort all
writers, we can iterate over all transactions and inspect their
statements. To abort all readers, MVCC trackers can be used. When MVCC
is disabled, we don't abort any transactions since memtx ones cannot yield.
Note that we swap indexes on DDL before the old space is invalidated, so
all index read gaps are actually moved to the new space. It's actually
incorrect but didn't break anything - we delete all transaction's
trackers when it is prepared or aborted, and we used to abort all
in-progress transactions on DDL. But with the new approach, we need
those gaps in the old space in order to find all space readers, so the
commit swaps read gaps back when moving space indexes.

A primary memtx difference is that it contains all system spaces, and
any schema change is a replace in a system space, hence, in memtx. So
the question is whether we should abort any active transactions on
schema changes that don't affect any space directly (for example, bump
of schema version on upgrade, replace in space `_schema`). The answer is
we shouldn't - all resources used by spaces are pinned, so we cannot
change, for example, a collation when it is used by a space, so let's
abort concurrent transactions only when a space is invalidated - it
happens on DDL operations directly affecting spaces (`space:format(...)`,
`space:create_index(...)` and so on).

Note that despite the commit populates `point_hole_item` with a new
member, its size remains the same - the object has a 8-byte alignment
and because of the `bool` in the end 7 bytes are wasted, so the new
member is put right to the end right before the boolean to use the space
used for alignment.

Closes #10377

NO_DOC=not documentable feature

(cherry picked from commit 3ac35eaeb38e416106aca24e43026e2dcf4d2fc8)