salad: reserve extents for matras_touch on BPS tree operations
Prior to this patch it was possible for the BPS tree to crash if one of
`matras_touch` calls is failed: the result of the function call was not
checked, because it was assumed the amount of extents we reserve prior
to an operation in the memtx engine will be enough for the functions.

Let's make the tree self-sufficient and reserve memory for `matras_touch`
calls prior to insertion and deletion.

NO_DOC=internal
NO_TEST=internal
NO_CHANGELOG=internal