replication: do not account booting replicas in sync quorum
With `bootstrap_strategy ~= 'legacy'` all replicas to which a node
managed to connect enter the pool of nodes with which the node has to
sync before leaving orphan mode.

This leads to inability to add new replicas to the cluster with
centralized configuration: if master manages to connect to some of the
new replicas before they are joined, it considers them part of sync
quorum and enters orphan mode. The nodes cannot be joined, because
master is read only, and master remains read only, because it can't sync
with a not yet booted node.

Let's fix this and not account connected but not yet booted nodes in
sync quorum.

Also fix the gh_10087 and gh_10088 tests. They started failing after the
patch because one of the replicas started becoming writable earlier,
before the other replica was bootstrapped. This made the other replica
occasionally bootstrap from it instead of bootstrapping from master.
Such situation was possible even before the patch, if one of the
replicas timed out waiting for the other to start, but was really
unlikely to happen, so the test wasn't ready for it.

Closes #10760

NO_DOC=bugfix

(cherry picked from commit 89cd7f43fd37e7afa459bf1a8f0e5159085c73ae)