config: add failover.log option options
Needed for EE tarantool/tarantool-ee#1016.

NO_CHANGELOG=changelog is a part of tarantool/tarantool-ee#1016
NO_TEST=will be tested as a part of tarantool/tarantool-ee#1016

Closes #10856

@TarantoolBot document
Title: New configuration options `failover.log`

Now `failover.log.to` and `failover.log.file` options can be used to
configure failover logger output.

Those options behave in similar way to `log.to` and `log.file` options.

`failover.log.to` defines the location for failover logs. This option
accepts the following values:
- `stderr`: write logs to the standard error stream.
- `file`: write logs to a file defined in `failover.log.file`.

`failover.log.file` specifies a file for logs destination. It is
mandatory to set this option if `failover.log.to` is set to `file`.
If `failover.log.to` is set to anything else, this option is ignored.

The option is enterprise-only since it's a part of the supervised
failover coordinator.