test: parallelize config-luatest/compat_test
It spawns a lot of processes and sometimes hits a test timeout limit on
my machine, when run in parallel with other tests.

Also, if I run the single test with the parallelization, the parallel
execution reduces time from 45 to 25 seconds.

NO_DOC=test adjustment
NO_CHANGELOG=see NO_DOC