config: allow supervised/native bs with coordinator
Previously, only `auto` bootstrap strategy was allowed to use with the
failover coordinator (when `replication.failover` = `supervised`). Now
we're going to support `supervised` and `native` strategies in the
coordinator. Let's allow to configure such a setup.

Part of #10857

NO_DOC=will be added together with the support of this setup in the
       coordinator
NO_CHANGELOG=same