config: don't fail on lack of sharding.roles
Before this change an instance with enabled sharding role assumes that
all the other instances in the cluster have some sharding role enabled
or at least an empty list of sharding roles.

This is a bug: a cluster may have instances that have no role in regards
to sharding.

Fixes #10458

NO_DOC=bugfix