lua/utils: `funcutils` for higher-order functions
This patch adds a `funcutils` module, which provides utility functions
for working with higher-order functions and callable objects.

The `chain2` function generates a new function that calls two given
functions in sequence. If one of the functions is `nil`, the other is
returned as is. If both are `nil`, the result is `nil`.

Needed for #10691

NO_DOC=internal
NO_CHANGELOG=internal
NO_TEST=internal