say: inherit log level when printing rate limit warning
`say_info_ratelimited()` logs a warning when it suppresses some
messages. This is confusing: we must not alert the user in this
case. Let's use the "info" log level instead of "warn".

This removes warnings logged on checkpoint and recovery when
Tarantool processes snapshot rows at a high rate.

NO_DOC=bug fix
NO_TEST=logging