test/fuzz: support txn_isolation option
The option `txn_isolation` was introduced in commit
ec750af66d94 ("txm: introduce transaction isolation levels")
and described in documentation (see [1] and [2]), but missed
in a page with `box.cfg` parameters [3]. It is a reason why it was
not used in the test `test_engine.lua` from the beginning.
The patch fixes that.

Note that isolation level 'linearizable' makes sense when space
is synchronous, the test uses only local spaces, so 'linearizable'
is not used.

1. https://www.tarantool.io/en/doc/latest/platform/atomic/txn_mode_mvcc/
2. https://www.tarantool.io/en/doc/latest/platform/atomic/transaction_model/
3. https://github.com/tarantool/doc/issues/4640

NO_CHANGELOG=testing
NO_DOC=testing
NO_TEST=testing