ci: run perf_micro workflow on performance label
The workflow `perf_micro` always executes on regular runners.
These runners usually are not suitable for performance testing
because benchmark results may vary significantly from run to run.

To be able to detect performance regressions before merging we
should:

1. Execute benchmarks on a special machine, where reference runs
   were done. This will make results comparable with each other.
2. Post obtained results to Grafana for collecting historical data
   and visualizing them.

The machine for running benchmarks is a valuable resource, and it
can quickly become a bottleneck if you run `perf_micro` workflow
on every push like we do with other workflows. To avoid this, the
patch uses a label 'performance', `perf_micro` workflow will
be started on a special machine only when 'performance' label is
set.

The second item requires access to a secret token that is not
available in workflows executed in the pull request. To workaround
this after the patch workflow use the event `pull_request_target`
that runs in the context of the base of the pull request, rather
than in the context of the merge commit, as the `pull_request`
event does. Running workflow in the context of the base allows to
use secrets defined in the main repository.

IMPORTANT: The `pull_request_target` event grants workflows
triggered by pull requests from forks access to repository secrets
and a read/write GITHUB_TOKEN. For further details see [1], [2] and
[3].

1. https://docs.github.com/en/actions/writing-workflows/choosing-when-your-workflow-runs/events-that-trigger-workflows#pull_request_target
2. https://github.blog/news-insights/product-news/github-actions-improvements-for-fork-and-pull-request-workflows/
3. https://securitylab.github.com/resources/github-actions-preventing-pwn-requests/

NO_CHANGELOG=performance
NO_DOC=performance
NO_TEST=performance

(cherry picked from commit 8de85f5c15f0092d99cf799a8bfe1f3c7f8293ba)