vinyl: abort readers on DDL
When a space is altered, we must abort all transactions writing to
the space because otherwise they would access a stale pointer stored in
`txn_stmt::space` on commit. We do that in `vinyl_space_invalidate()`,
which is called when a space is removed from the cache. There we iterate
over all writing transactions and abort those of them whose write sets
contain statements for the altered space, see commit c602cc2c82963
("vinyl: abort rw transactions before DDL"). To handle transactions
that are currently waiting for a disk read before they can create
a write statement (such as an update, which needs to read the old
tuple first), we also store the current space when we begin a statement,
see commit 4f62ec2102f49 ("vinyl: make tx_manager_abort_writers_for_ddl
more thorough").

However, those checks actually aren't enough: a transaction statement
may turn out to be a no-op, for example, if it updates a non-existent
key, in which case it isn't added to the transaction write set. If we
don't abort such a statement, we'll get a use-after-free bug when we
try to commit it. Let's handle this issue by aborting all transactions
that read from the altered space. This would also be consistent with
the memtx engine.

Closes #10707
Closes #10786

NO_DOC=bug fix